# internal/httpserver/swaggerdocs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "GoDNS API",
        "contact": {
            "name": "GoDNS Support",
            "url": "https://github.com/rogerwesterbo/godns",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/rogerwesterbo/godns/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:14000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/zones/{zone}/records": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Add a new DNS record to an existing zone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Records"
                ],
                "summary": "Create a DNS record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Record to create",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSRecord"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Record created",
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSRecord"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Zone not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Record already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/zones/{zone}/records/{name}/{type}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get a specific DNS record by name and type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Records"
                ],
                "summary": "Get a DNS record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record name (e.g., www.example.lan.)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record type (e.g., A, AAAA, CNAME)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record details",
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSRecord"
                        }
                    },
                    "404": {
                        "description": "Zone or record not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Update an existing DNS record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Records"
                ],
                "summary": "Update a DNS record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record name (e.g., www.example.lan.)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record type (e.g., A, AAAA, CNAME)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated record data",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSRecord"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Zone or record not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Delete a specific DNS record from a zone",
                "tags": [
                    "Records"
                ],
                "summary": "Delete a DNS record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record name (e.g., www.example.lan.)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record type (e.g., A, AAAA, CNAME)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Record deleted"
                    },
                    "404": {
                        "description": "Zone or record not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_rogerwesterbo_godns_internal_models.DNSRecord": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Fully qualified domain name",
                    "type": "string",
                    "example": "www.example.lan."
                },
                "ttl": {
                    "description": "Time to live in seconds",
                    "type": "integer",
                    "example": 300
                },
                "type": {
                    "description": "A, AAAA, MX, NS, TXT, etc.",
                    "type": "string",
                    "example": "A"
                },
                "value": {
                    "description": "The record value (IP, hostname, etc.)",
                    "type": "string",
                    "example": "192.168.1.100"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:14101/realms/godns/protocol/openid-connect/token"
        }
    },
...
}